+++ York GC Members App/MHSystems/app/src/main/java/com/mh/systems/york/ui/activites/CompletedDetailActivity.java
@@ -327,7 +327,7 @@ public class CompletedDetailActivity extends BaseActivity {
 
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
+            reportRollBarException(CompetitionDetailActivity.class.getSimpleName(), e.toString());
         }
         hideProgress();
     }