+++ York GC Members App/MHSystems/app/src/main/java/com/mh/systems/york/ui/activites/EditDetailsActivity.java
@@ -294,7 +294,7 @@ public class EditDetailsActivity extends BaseActivity implements View.OnClickLis
 
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
+            reportRollBarException(EditDetailsActivity.class.getSimpleName(), e.toString());
         }
         hideProgress();
     }
@@ -577,7 +577,7 @@ public class EditDetailsActivity extends BaseActivity implements View.OnClickLis
         } catch (Exception e) {
             hideProgress();
             Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
+            reportRollBarException(EditDetailsActivity.class.getSimpleName(), e.toString());
             llEditDetailGroup.setVisibility(View.GONE);
         }
     }