+++ York GC Members App/MHSystems/app/src/main/java/com/mh/systems/york/ui/activites/ForgotPasswordActivity.java
@@ -242,7 +242,7 @@ public class ForgotPasswordActivity extends BaseActivity implements View.OnClick
             hideProgress();
         } catch (Exception e) {
             hideProgress();
-            Log.e(LOG_TAG, "" + e.getMessage());
+            reportRollBarException(ForgotPasswordActivity.class.getSimpleName(), e.toString());
         }
     }