+++ York GC Members App/MHSystems/app/src/main/java/com/mh/systems/york/ui/activites/TopUpActivity.java
@@ -193,7 +193,7 @@ public class TopUpActivity extends BaseActivity {
         super.onActivityResult(requestCode, resultCode, data);
 
         /**
-         * If PAYMENT status SUCCESS then navigate user to {@link com.mh.systems.sunningdale.fragments.FinanceFragment}
+         * If PAYMENT status SUCCESS then navigate user to {@link com.mh.systems.york.fragments.FinanceFragment}
          * Otherwise, retain on this screen.
          */
         if (resultCode == ACTION_MAKE_PAYMENT) {
@@ -202,8 +202,9 @@ public class TopUpActivity extends BaseActivity {
                 if (isPaymentSuccess) {
                     onBackPressed();
                 }
-            } catch (Exception exp) {
-                Log.e(LOG_TAG, "" + exp.toString());
+            } catch (Exception e) {
+                Log.e(LOG_TAG, "" + e.toString());
+                reportRollBarException(TopUpActivity.class.getSimpleName(), e.toString());
             }
         }
     }
@@ -464,7 +465,7 @@ public class TopUpActivity extends BaseActivity {
             hideProgress();
         } catch (Exception e) {
             hideProgress();
-            Log.e(LOG_TAG, "" + e.getMessage());
+            reportRollBarException(TopUpActivity.class.getSimpleName(), e.toString());
             showNoTopUpView(false);
         }
     }