+++ York GC Members App/MHSystems/app/src/main/java/com/mh/systems/york/ui/fragments/HandicapFragment.java
@@ -30,6 +30,7 @@ import com.google.gson.JsonObject;
 import com.mh.systems.york.R;
 import com.mh.systems.york.ui.activites.BaseActivity;
 import com.mh.systems.york.ui.activites.HCapHistoryActivity;
+import com.mh.systems.york.ui.activites.MembersActivity;
 import com.mh.systems.york.ui.activites.ShowCertificateWebview;
 import com.mh.systems.york.ui.activites.YourAccountActivity;
 import com.mh.systems.york.utils.constants.ApplicationGlobal;
@@ -373,7 +374,7 @@ public class HandicapFragment extends Fragment implements OnChartValueSelectedLi
             ((BaseActivity) getActivity()).hideProgress();
         } catch (Exception e) {
             Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
+            ((YourAccountActivity)getActivity()).reportRollBarException(HandicapFragment.class.getSimpleName(), e.toString());
             ((BaseActivity) getActivity()).hideProgress();
         }
     }