+++ York GC Members App/MHSystems/app/src/main/java/com/mh/systems/york/ui/fragments/MyDetailsFragment.java
@@ -13,6 +13,7 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.google.gson.JsonObject;
+import com.mh.systems.york.ui.activites.MembersActivity;
 import com.mh.systems.york.ui.activites.YourAccountActivity;
 import com.newrelic.com.google.gson.Gson;
 import com.newrelic.com.google.gson.reflect.TypeToken;
@@ -110,6 +111,8 @@ public class MyDetailsFragment extends Fragment {
             ((YourAccountActivity) getActivity()).setiBalanceType(0);
             /* ++++++++++++++++  END OF PURSE API FEATURE ++++++++++++++++ */
 
+            YourAccountActivity.isRefreshEnable = true;
+
             /**
              *  Check internet connection before hitting server request.
              */
@@ -200,7 +203,7 @@ public class MyDetailsFragment extends Fragment {
         } catch (Exception e) {
             ((BaseActivity) getActivity()).hideProgress();
             Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
+            ((YourAccountActivity)getActivity()).reportRollBarException(MyDetailsFragment.class.getSimpleName(), e.toString());
             llMyDetailGroup.setVisibility(View.GONE);
         }
     }